Added "delete temp directory tree" and sort cell rows.

+++ dpaviser-jhove-report-main/src/main/java/dk/statsbiblioteket/dpaviser/report/Main.java
@@ -1,5 +1,6 @@
 package dk.statsbiblioteket.dpaviser.report;
 
+import dk.statsbiblioteket.dpaviser.report.helpers.CellRowComparator;
 import dk.statsbiblioteket.dpaviser.report.helpers.POIHelpers;
 import dk.statsbiblioteket.dpaviser.report.helpers.XPathHelpers;
 import dk.statsbiblioteket.dpaviser.report.jhove.JHoveHelpers;
@@ -9,10 +10,10 @@ import java.io.File;
 import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.OutputStream;
-import java.nio.file.DirectoryNotEmptyException;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
+import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 import java.util.function.Predicate;
@@ -24,11 +25,15 @@ public class Main {
 
     public static void main(String[] args) throws Exception {
         if (args.length < 1) {
-            System.err.println("Usage: java " + Main.class.getName() + " spreadsheet.xml infomedia-dump-dir");
+            System.err.println("Usage: java " + Main.class.getName() + " spreadsheet.xsl infomedia-dump-dir");
         }
         long start = System.currentTimeMillis();
 
         File tmpDir = com.google.common.io.Files.createTempDir();
+        System.out.println(tmpDir.getAbsolutePath());
+        new File(tmpDir, "a").mkdir();
+        new File(tmpDir, "a/b").mkdir();
+        new File(tmpDir, "c").mkdir();
 
         Pattern singleEditionDirPattern = Pattern.compile("^.*/[A-Z0-9]+/[0-9]{4}/(0[1-9]|1[0-2])/(0[1-9]|[12][0-9]|3[01])$");
 
@@ -49,40 +54,47 @@ public class Main {
 
         System.out.println(cellRows.size() + " rows.");
 
-        Collections.sort(cellRows, (left, right) -> {
-            for (int i = 0; i < Math.min(left.size(), right.size()); i++) {
-                int comparison = left.get(i).compareTo(right.get(i));
-                if (comparison != 0) {
-                    return comparison;
-                }
-            }
-            // one list exhausted without a difference - the shortest is "before" the longest.
-            return right.size() - left.size();
-        });
+        Collections.sort(cellRows, new CellRowComparator());
 
         try (OutputStream out = new FileOutputStream(args[0])) {
             POIHelpers.workbookFor(cellRows).write(out);
         }
 
-        // if we get this far, clean up temp dir.  http://stackoverflow.com/a/5039900/53897
-        Path dirPath = tmpDir.toPath();
-        Files.walk(dirPath)
-                .filter(Files::isRegularFile)
-                .forEach((path) -> {
-                    try {
-                        Files.delete(path);
-                    } catch (IOException e) {
-                        e.printStackTrace();
-                    }
-                });
-        // For now do not delete subdirectories (need to delete the tree from the bottom).
-        try {
-            Files.delete(dirPath);
-        } catch (DirectoryNotEmptyException e) {
-            System.out.println("Directory not empty - " + e);
-        }
+        deleteTreeBelowPath(tmpDir.toPath());
 
         System.out.println(System.currentTimeMillis() - start + " ms.");
     }
 
+    /**
+     * Delete the whole file tree below the given path.  Files first while traversing, and the directories at the end.
+     */
+    private static void deleteTreeBelowPath(Path startHerePath) throws IOException {
+        List<Path> dirPaths = new ArrayList<>();
+
+        // Delete files as we see them.  Save directories for deletion later.
+        Files.walk(startHerePath)
+                .forEach(path -> {
+                    if (Files.isDirectory(path)) {
+                        dirPaths.add(path);
+                    } else {
+                        try {
+                            Files.delete(path);
+                        } catch (IOException e) {
+                            e.printStackTrace(); // so we can continue even if a file is open.
+                        }
+                    }
+                });
+
+        Collections.sort(dirPaths);
+        Collections.reverse(dirPaths);
+
+        for (Path path : dirPaths) {
+            try {
+                //System.out.println(path);
+                Files.delete(path);
+            } catch (IOException e) {
+                System.out.println("Could not delete " + path + ": " + e); // better suggestions?
+            }
+        }
+    }
 }
